+++ tests/test_evaluate.py
@@ -3,7 +3,7 @@ import pytest  # noqa: F401
 from query_builder_widget import evaluate
 from query_builder_widget.types import QueryGroup
 
-CONDITIONS: QueryGroup = {
+GOOD_CONDITIONS: QueryGroup = {
     "rules": [
         {"field": "server", "value": "1255250024741212262", "operator": "equals"},
         {
@@ -17,12 +17,29 @@ CONDITIONS: QueryGroup = {
     "condition": "OR",
 }
 
+IGNORABLE_CONDITIONS_AND: QueryGroup = {"rules": [], "condition": "AND"}
+IGNORABLE_CONDITIONS_OR: QueryGroup = {"rules": [], "condition": "OR"}
 
-async def test_evaluation():
+
+async def test_evaluation_good():
     data = {"server": "1255250024741212262", "completion": "50", "ballcount": "7"}
 
     result = await evaluate(
-        CONDITIONS, data, field_labels={"server": "Server", "completion": "Ball completion", "ballcount": "Ball count"}
+        GOOD_CONDITIONS,
+        data,
+        field_labels={"server": "Server", "completion": "Ball completion", "ballcount": "Ball count"},
     )
+
+    assert result.passed, ", ".join(result.failures)
+
+
+async def test_evaluation_ignorable_and():
+    result = await evaluate(IGNORABLE_CONDITIONS_AND, {})
+
+    assert result.passed, ", ".join(result.failures)
+
+
+async def test_evaluation_ignorable_or():
+    result = await evaluate(IGNORABLE_CONDITIONS_OR, {})
 
     assert result.passed, ", ".join(result.failures)