+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.1] - 2026-09-14
+
+### Added
+
+- Added `test_evaluation_ignorable_and` and `test_evaluation_ignorable_or` tests.
+
+### Fixed
+
+- Fixed evaluation failing when given a query group with empty rules and an `OR` condition.
+
 ## [1.2.0] - 2026-09-12
 
 ### Added